+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Change Log
 
 ### master (untagged)
 
+* Add AlexNet example
+
 ### 0.0.1 (2017-06-23)
 
 * Add toy tutorial